% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logKDE.R
\name{bw.logCV}
\alias{bw.logCV}
\title{Computes least squares CV bandwidth for log domain KDE.}
\usage{
bw.logCV(y, grid = 21, NB = 512)
}
\arguments{
\item{y}{the data from which the estimate is to be computed.}

\item{grid}{number of points for bw grid}

\item{NB}{number of points to estimate the KDE at during the CV loop}
}
\value{
bw the optimal CV bandwidth
}
\description{
Computes least squares CV bandwidth for log domain KDE.
}
\examples{
bw.logCV(rchisq(100,10), grid=21, NB=512)

}
