% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logKDE.R
\name{bw.logG}
\alias{bw.logG}
\title{Computes least squares bandwidth for log domain KDE using modified silverman rule.}
\usage{
bw.logG(x)
}
\arguments{
\item{x}{the data from which the estimate is to be computed.}
}
\value{
h the optimal bandwidth
}
\description{
Computes least squares bandwidth for log domain KDE using modified silverman rule.
}
\examples{
bw.logG(rchisq(100,10))

}
