% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logKDE.R
\name{logdensity_fft}
\alias{logdensity_fft}
\title{Computes Kernel Density Estimates in Log Domain using FFT}
\usage{
logdensity_fft(x, bw = "nrd0", adjust = 1, kernel = "gaussian",
  weights = NULL, give.Rkern = FALSE, n = 512, from, to, cut = log(3),
  na.rm = FALSE)
}
\arguments{
\item{x}{the data from which the estimate is to be computed.}

\item{bw}{bandwidth.}

\item{adjust}{the bandwidth used is actually adjust*bw.}

\item{kernel}{choose from "gaussian", "epanechnikov", "triangular", "uniform", "laplace" and "logistic".}

\item{weights}{numeric vector of non-negative observation weights.}

\item{give.Rkern}{honesttly dont know what this does}

\item{n}{the number of equally spaced points at which the density is to be estimated.}

\item{from, to}{the left and right-most points of the grid at which the density is to be estimated; the defaults are cut * bw outside of range(x).}

\item{cut}{by default, the values of from and to are cut bandwidths beyond the extremes of the data}

\item{na.rm}{standard deal}
}
\value{
Density object. See help(density).
}
\description{
Computes Kernel Density Estimates in Log Domain using FFT
}
\examples{
logdensity_fft(abs(rnorm(100)), from =0.01, to= 2.5, kernel = 'logistic')

}
