% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logKDE.R
\name{bw.logG}
\alias{bw.logG}
\title{Bandwidth estimation for strictly positive distributions.}
\usage{
bw.logG(x)
}
\arguments{
\item{x}{numeric vector of the data. Must be strictly positive, will be log transformed during estimation.}
}
\value{
bw the optimal bandwidth.
}
\description{
Computes bandwidth for log domain KDE using the Silverman rule.
}
\examples{
bw.logG(rchisq(100,10))

}
